requirements.txt
Django~=3.1
djangorestframework~=3.12.4
beautifulsoup4~=4.10.0
soupsieve~=2.3.1
pip~=21.2.3
pytz~=2021.3
psycopg2~=2.9.2
asgiref~=3.2.10
MarkupSafe~=2.0.1
certifi~=2021.10.8
itypes~=1.2.0
coreapi~=2.3.3
coreschema~=0.0.4
drf_yasg~=1.20
requests~=2.26.0
uritemplate~=4.1.1
urllib3~=1.26.7
inflection~=0.5.1
idna~=3.3
pyparsing~=3.0.6
setuptools~=57.4.0